Pull cli utilites out of the generic utilities file
