+++ tests/lib/io/test_data_root.py
